changed data type double to float in migration

+++ db/migrate/20111025020244_change_score_type_from_integer_to_double_at_ratings.rb
@@ -1,6 +1,6 @@
 class ChangeScoreTypeFromIntegerToDoubleAtRatings < ActiveRecord::Migration
   def up
-    change_column :ratings, :score, :double
+    change_column :ratings, :score, :float
   end
 
   def down
