+++ db/migrate/20111028001327_add_average_score_to_article.rb
@@ -1,5 +1,5 @@
 class AddAverageScoreToArticle < ActiveRecord::Migration
   def change
-    add_column :articles, :average_score, :double, :default => 0
+    add_column :articles, :average_score, :float, :default => 0
   end
 end
